@@ -1,5 +1,5 @@
 import { Http, Headers, RequestOptions, Response } from '@angular/http';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { BehaviorSubject } from 'rxjs';
 import { Router } from '@angular/router';
 export class AppService {
     private messageSource;
